added docker-compose and moved tools and made dockerfile
