+++ backend/app.py
@@ -1,7 +1,7 @@
 from typing import Union
 from fastapi import FastAPI
 # from agents.image_extractor_from_files import get_file
-from agents import image_extractor_from_files
+from backend.agents import image_extractor_from_files
 from pydantic import BaseModel
 from fastapi import UploadFile, File, HTTPException
 import tempfile
@@ -16,13 +16,12 @@ class FileURL(BaseModel):
 def read_root():
     return {"Hello": "World"}
 
-
 @app.get("/items/{item_id}")
 def read_item(item_id: int, q: Union[str, None] = None):
     return {"item_id": item_id, "q": q}
 
-@app.post("/github-file")
-async def receive_github_file(payload: FileURL):
+@app.post("/url-file")
+async def receive_url_file(payload: FileURL):
     await image_extractor_from_files.get_file(payload.url)
     # Add logic to process the file at payload.url
     return {"status": "File received", "file_url": payload.url}
@@ -50,6 +49,7 @@ async def upload_file(file: UploadFile = File(...)):
     return {"file_name": file.filename, "status": "Uploaded successfully"}
 
 
-if __name__ == "__main__":
-    import uvicorn
-    uvicorn.run(app, host="127.0.0.1", port=8000)
+# if __name__ == "__main__":
+#     import uvicorn
+#     uvicorn.run(app, host="127.0.0.1", port=3100)
+
