+++ containers/image_extractor_dockerfile
@@ -1,0 +1,15 @@
+FROM python:3.12-slim
+
+WORKDIR /code
+
+COPY ../backend ./backend
+
+COPY ../requirements.txt .
+
+RUN pip install --no-cache-dir -r requirements.txt
+
+EXPOSE 3100
+
+# CMD [ "fastapi", "run", "backend/app.py --port 3100" ]
+
+CMD ["gunicorn", "-w", "4", "-k", "uvicorn.workers.UvicornWorker", "backend.app:app", "--bind", "0.0.0.0:3100"]
