+++ docker-compose.yaml
@@ -1,0 +1,11 @@
+services:
+  app:
+    build:
+      context: .
+      dockerfile: containers/image_extractor_dockerfile
+    env_file:
+      - .env
+    ports:
+      - "8000:3100"
+    # volumes:
+    #   - ../backend:/app/backend
