+++ environment.yaml
@@ -49,6 +49,7 @@ dependencies:
       - et-xmlfile==2.0.0
       - fastapi==0.115.8
       - frozenlist==1.5.0
+      - gunicorn==23.0.0
       - h11==0.14.0
       - httpcore==1.0.7
       - httpx==0.28.1
@@ -93,6 +94,7 @@ dependencies:
       - pymeta3==0.5.1
       - python-docx==1.1.2
       - python-dotenv==1.0.1
+      - python-multipart
       - python-pptx==1.0.2
       - pyyaml==6.0.2
       - referencing==0.36.2
