+++ requirements.txt
@@ -19,6 +19,7 @@ distro==1.9.0
 et_xmlfile==2.0.0
 frozenlist==1.5.0
 fastapi==0.115.8
+gunicorn
 h11==0.14.0
 httpcore==1.0.7
 httpx==0.28.1
@@ -64,6 +65,7 @@ PyMeta3==0.5.1
 python-docx==1.1.2
 python-dotenv==1.0.1
 python-pptx==1.0.2
+python-multipart
 PyYAML==6.0.2
 referencing==0.36.2
 requests==2.32.3
